refactor: extract recording status poll from App
