+++ src/app/App.tsx
@@ -31,7 +31,6 @@ import {
   dictationHelperCommand,
   downloadNoteAudio,
   ensureHermesBridgeSession,
-  getRecordingStatus,
   getNote,
   LIVE_TRANSCRIPT_EVENT,
   listCompletedSessions,
@@ -117,11 +116,7 @@ import {
 import { type MaxUpgradeTransport } from "../lib/billing-actions";
 import { type MaxGrantWait } from "../lib/max-upgrade";
 import { reconcileToStable, relaunchJune, type JuneUpdate } from "../lib/updater";
-import {
-  PROCESSING_DEMO_NOTE_ID,
-  RECORD_NOTICES_DEMO_SESSION_ID,
-  shouldPollProcessingStatus,
-} from "./processing-polling";
+import { PROCESSING_DEMO_NOTE_ID, shouldPollProcessingStatus } from "./processing-polling";
 import { attachScrollThumbFade } from "../lib/scroll-thumb-fade";
 import { createInitialState, notesReducer } from "./state/app-state";
 import {
@@ -135,7 +130,6 @@ import {
 import {
   handleTitlebarPointerDown,
   isAccessibilityBlocked,
-  isAppErrorCode,
   isCreateNoteShortcut,
   isMicrophoneRecordingBlocked,
   isNewSessionShortcut,
@@ -197,6 +191,8 @@ import { useAppBootstrap } from "./use-app-bootstrap";
 import { useAppTabEvents } from "./use-app-tab-events";
 
 import { useAppDevDemos } from "./use-app-dev-demos";
+
+import { useRecordingStatusPoll } from "./use-recording-status-poll";
 
 export function App() {
   const replayOnboarding = shouldReplayOnboarding();
@@ -1479,57 +1475,11 @@ export function App() {
     });
   }
 
-  useEffect(() => {
-    if (!state.recordingStatus || !["recording", "paused"].includes(state.recordingStatus.state)) {
-      return;
-    }
-    const sessionId = state.recordingStatus.sessionId;
-    // The dev __recordNoticesDemo session lives only in the reducer — there is
-    // no backend recording to poll, and getRecordingStatus would clear the
-    // synthetic bar with a "recording not found". Stripped from production via
-    // import.meta.env.DEV. See lib/record-notices-demo.ts.
-    if (import.meta.env.DEV && sessionId === RECORD_NOTICES_DEMO_SESSION_ID) {
-      return;
-    }
-    // Drops in-flight responses once this effect is torn down. Without it, a
-    // poll that was already in flight when the user hit stop resolves after
-    // recordingStatusCleared and resurrects the recorder bar with a stale
-    // status — and since polling for that state never restarts, the bar would
-    // be stuck on screen indefinitely.
-    let cancelled = false;
-    // ~20Hz so the waveform tracks speech as snappily as the dictation HUD
-    // (which is event-driven at ~25Hz). The polled equivalent for the recorder;
-    // each poll coalesces the peaks since the last one (see Waveform.tsx). Audio
-    // is sampled every ~5–10ms in Rust, so there's always a fresh peak waiting;
-    // 100ms left the bars a beat behind the voice.
-    let inFlight = false;
-    const interval = window.setInterval(() => {
-      if (inFlight) return;
-      inFlight = true;
-      getRecordingStatus(sessionId)
-        .then((status) => {
-          if (!cancelled) dispatch({ type: "recordingStatusChanged", status });
-        })
-        .catch((err: unknown) => {
-          if (cancelled) return;
-          if (isAppErrorCode(err, "recording_not_found")) {
-            // The backend no longer tracks this session — clear the bar
-            // instead of polling a dead session forever. The reducer ignores
-            // this if a newer session already replaced it.
-            dispatch({ type: "recordingSessionLost", sessionId });
-            return;
-          }
-          setError(messageFromError(err));
-        })
-        .finally(() => {
-          inFlight = false;
-        });
-    }, 50);
-    return () => {
-      cancelled = true;
-      window.clearInterval(interval);
-    };
-  }, [state.recordingStatus?.sessionId, state.recordingStatus?.state]);
+  useRecordingStatusPoll({
+    dispatch,
+    setError,
+    state,
+  });
 
   useEffect(() => {
     if (
